+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 - Row deletion is now executed in an SQL transaction. If the number of affected rows is not equal to
   1 (indicating a logic error in our code), then the transaction is rolled back.
 
+- A `Count rows` button is displayed in table row-list buffers.
+
 
 ## [0.7] - 2024-05-23